remove custom widgets implementation, and instead move code to graph_component
